move: message exchange directory
